pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "golfclip"
version = "0.1.0"
description = "Mac desktop app that transforms iPhone golf recordings into YouTube-ready clips"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Private"}
authors = [
    {name = "Eli Coon"}
]

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",

    # ML & Computer Vision
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ultralytics>=8.1.0",  # YOLOv8
    "opencv-python>=4.9.0",

    # Audio Analysis
    "librosa>=0.10.1",
    "soundfile>=0.12.1",

    # Video Processing
    "ffmpeg-python>=0.2.0",

    # Utilities
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "tqdm>=4.66.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
golfclip = "backend.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
